test migrated

+++ contract/contracts/TicketContract.sol
@@ -11,79 +11,69 @@ import "@openzeppelin/contracts/token/ERC1155/extensions/ERC1155Pausable.sol";
 /// @notice Allows an organizer to publish events, and other users to buy tickets, more for the future
 /// @custom:security-contact @javierlinked
 contract TicketContract is ERC1155, Ownable, ERC1155Pausable, ERC1155Burnable {
-
     /// @dev variariables and counters used by the contract and by the frontend, in order to crate a list
-    uint public nonce;
-    uint[] public tokenIds;
-    uint public tokenIdsLength;
-    
+    uint256 public nonce;
+    uint256[] public tokenIds;
+    uint256 public tokenIdsLength;
+
     struct Ticket {
-        uint id;
+        uint256 id;
         string name;
-        uint price;     // in gwei
-        uint maxSellPerPerson;  // max number of tickets that can be kept per person
+        uint256 price; // in gwei
+        uint256 maxSellPerPerson; // max number of tickets that can be kept per person
         string infoUrl; // this is to be removed and kept off-chain
     }
 
-    mapping(uint => Ticket) public tickets;
+    mapping(uint256 => Ticket) public tickets;
 
     /// @notice This event is emmited when a ticket is created
     /// @param id the id of the ticket
     /// @param minter address of the minter
     /// @param name name of the ticket
-    /// @param price price of the ticket 
+    /// @param price price of the ticket
     /// @param amount total supply of tickets
     /// @param maxSellPerPerson the maximum number of tickets that can be hold per person
     event TicketCreated(
-        uint indexed id,
+        uint256 indexed id,
         address indexed minter,
         string indexed name,
-        uint price,
-        uint amount,
-        uint maxSellPerPerson
+        uint256 price,
+        uint256 amount,
+        uint256 maxSellPerPerson
     );
 
-    //// @notice this event is emmited when a ticket is sold
+    /// @notice this event is emmited when a ticket is sold
     /// @param id the id of the ticket
     /// @param seller the address of the seller
     /// @param buyer the address of the buyer
     /// @param amount the amount of tickets sold
-    event TicketSold(
-        uint indexed id,
-        address indexed seller,
-        address indexed buyer,
-        uint amount
-    );
+    event TicketSold(uint256 indexed id, address indexed seller, address indexed buyer, uint256 amount);
 
     /// @notice verifies if amount is valid
-    modifier validAmount(uint amount, uint maxSellPerPerson) {
+    modifier validAmount(uint256 amount, uint256 maxSellPerPerson) {
         require(amount > 0, "Incorrect amount");
         require(maxSellPerPerson > 0, "Incorrect maxSellPerPerson");
         require(amount >= maxSellPerPerson, "Incorrect amount");
         _;
     }
 
-
     /// @notice this verifies the payment is exact or cancels the transaction
-    modifier paidExactly(uint amount, uint id) {
-        uint total = amount * tickets[id].price;
+    modifier paidExactly(uint256 amount, uint256 id) {
+        uint256 total = amount * tickets[id].price;
         require(total == msg.value, "Incorrect amount");
         _;
     }
 
     /// @notice this verifies the minter has enough tickets to sell
-    modifier allowedSell(uint amount, uint id) {
+    modifier allowedSell(uint256 amount, uint256 id) {
         require(amount > 0, "Amount must bigger than 0");
         require(balanceOf(owner(), id) >= amount, "Not enough tickets");
-        uint buying = balanceOf(msg.sender, id) + amount;
-        require(
-            buying <= tickets[id].maxSellPerPerson,
-            "Max ammount per person reached"
-        );
+        uint256 buying = balanceOf(msg.sender, id) + amount;
+        require(buying <= tickets[id].maxSellPerPerson, "Max ammount per person reached");
         _;
     }
 
-    modifier validate(uint id) {
+    modifier validate(uint256 id) {
         require(tickets[id].price != 0, "Ticket does not exist");
         _;
     }
@@ -96,20 +86,20 @@ contract TicketContract is ERC1155, Ownable, ERC1155Pausable, ERC1155Burnable {
     /// @notice Crates a ticket
     /// @dev After validations and internal variable set, delegates to the ERC1155 contract and emmits the event
     /// @param name name of the ticket
-    /// @param price price of the ticket 
+    /// @param price price of the ticket
     /// @param amount total supply of tickets
     /// @param maxSellPerPerson the maximum number of tickets that can be hold per person
     /// @param infoUrl extra info about the ticket
     /// @param data data required by ERC1155
     function create(
         string memory name,
-        uint price,
-        uint amount,
-        uint maxSellPerPerson,
+        uint256 price,
+        uint256 amount,
+        uint256 maxSellPerPerson,
         string memory infoUrl,
         bytes memory data
     ) public onlyOwner validAmount(amount, maxSellPerPerson) whenNotPaused {
-        uint newId = ++nonce;
+        uint256 newId = ++nonce;
 
         address owner = msg.sender;
         Ticket memory newTicket = Ticket(newId, name, price, maxSellPerPerson, infoUrl);
@@ -127,20 +117,13 @@ contract TicketContract is ERC1155, Ownable, ERC1155Pausable, ERC1155Burnable {
     /// @param amount amount of tickets to sell
     /// @param data data required by ERC1155
     function buy(
-        uint id,
-        uint amount,
+        uint256 id,
+        uint256 amount,
         bytes memory data
-    )
-        public
-        payable
-        allowedSell(amount, id)
-        paidExactly(amount, id)
-        validate(id)
-        whenNotPaused
-    {
+    ) public payable allowedSell(amount, id) paidExactly(amount, id) validate(id) whenNotPaused {
         address payable owner = payable(owner());
         address buyer = msg.sender;
-        (bool sent, ) = owner.call{value: msg.value}("");
+        (bool sent, ) = owner.call{ value: msg.value }("");
         require(sent, "Failed to send Ether");
         _safeTransferFrom(owner, buyer, id, amount, data);
         emit TicketSold(id, owner, buyer, amount);
@@ -150,10 +133,9 @@ contract TicketContract is ERC1155, Ownable, ERC1155Pausable, ERC1155Burnable {
     /// @dev after verifications delegates to the ERC1155 burn of the token
     function burn(
         address account,
-        uint id,
-        uint amount
-    ) public override
-    whenNotPaused {
+        uint256 id,
+        uint256 amount
+    ) public override whenNotPaused {
         // NOTE:  to myself, not related to _balances[][] but to existence of tokens in the contract
         _burn(account, id, amount);
     }
